// WeatherApp/src/Interface.tsx
export interface IWeatherData {
    base: string;
    cloud: {
      all: number;
    };
    cod: number;
    coord: {
      lat: number;
      lon: number;
    };
    dt: number;
    id: number;
    main: {
      feels_like: number;
      grnd_level: number;
      humidity: number;
      pressure: number;
      sea_level: number;
      temp: number | undefined;
      temp_max: number;
      temp_min: number;
    };
    name: string;
    sys: {
      country: string;
      sunrise: number;
      sunset: number;
    };
    timezone: number;
    visibility: number;
    weather: [
      {
        description: string;
        icon: string;
        id: number;
        main: string;
      }
    ];
    wind: {
      deg: number;
      gust: number;
      speed: number;
    };
  }